requirements.txt
streamlit
yfinance
plotly
pandas
numpy
supabase
python-dotenv
datetime